fixed macOS issue, and switches issue. Simplified login
Changed readme to instruct macOS users to install requests if they run into issues.

Also clarified how to alter code for login/country/application (from the perspective of a noob)

In the code, encourages users to directly type in their login instead of working with environments - more user friendly? I was unsure of how to use it as a python noob.

Also added Matt's suggestion of adding switches to the list of devices.